Common: Remove else's after if's that stop function executions.
If we have an `if`-`else` command where we never continue the execution from
the `if` block, then we can be free to unwrap the `else` body from the else.

This improves readability by reducing one level of indentation.

Note that there was one instance of this removal where the body of the
`else` was not even indented, which made reading the code confusing
(especially for those that program in languages like Python where
indentation matters).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>